Updated JSON data checking
I started creating the data validation checks needed to verify that the data in each field for the dogs is of the correct form, which requires the use of isinstance to trigger exceptions. In looking up information about isinstance, I found several sources, notably http://canonical.org/~kragen/isinstance/, that make the case that isinstance is poor programming form. I've now stopped developing in this direction.