Revised website and JavaScript for form recall
To test the use of the jQuery function .getJSON(), a form with styled radio buttons and a multiple selection list was added to the HTML, along with a field to be updated with the information from the JSON database. The script was updated with jQuery functions to capture the inputs to the fields on the website and pull the appropriate data from the JSON database.